Refactor to split out process_item_list from charts.py
