+++ template_reports/office_renderer/exceptions.py
@@ -26,9 +26,3 @@ class ChartError(Exception):
     """Raised when an error occurs while processing a chart."""
 
     pass
-
-
-class BadChartDataResultError(Exception):
-    """Raised when a processed value is invalid (i.e. result is not a float)."""
-
-    pass
