+++ template_reports/templating/exceptions.py
@@ -10,6 +10,12 @@ class MissingDataException(Exception):
     pass
 
 
+class BadFloatDataResultError(Exception):
+    """Raised when a processed value is invalid (i.e. result is not a float)."""
+
+    pass
+
+
 class BadTemplateModeError(Exception):
     """Raised when an invalid mode is passed to process_text."""
